pyproject.toml
[tool.poetry]
name = "genai-lab"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvd.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
httpx = "^0.27.2"
ollama = "^0.3.3"
loguru = "^0.7.2"
tqdm = "^4.66.6"
colorama = "^0.4.6"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
sentence-transformers = "^3.2.1"
transformers = { extras = ["sentencepiece"], version = "^4.46.1" }
openai = "^1.53.0"
python-dotenv = "^1.0.1"
langchain = "^0.3.7"
langchainhub = "^0.1.21"
langchain-openai = "^0.2.5"
langchain-ollama = "^0.2.0"
langchain-community = "^0.3.5"
langchain-chroma = "^0.1.4"
tiktoken = "^0.8.0"
dagster = "^1.9.0"
dagster-postgres = "^0.25.0"
chromadb-client = "^0.5.17"
unstructured = { extras = ["md"], version = "^0.16.4" }
diffusers = { extras = ["torch"], version = "^0.31.0" }
huggingface-hub = { extras = ["cli"], version = "^0.26.2" }
langsmith = "^0.1.146"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.0"
ipython = "^8.28.0"
jupyter = "^1.1.1"
notebook = "^7.2.2"
dagster-webserver = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dagster]
module_name = "pipeline.definitions"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B", "UP"]
ignore = ["E203", "E266", "E501", "E402"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query"]

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["pydantic.validator"]
